requirements.txt
alembic==1.0.5
astroid==2.1.0
autopep8==1.4.3
bcrypt==3.1.6
cffi==1.11.5
gunicorn==19.9.0
hupper==1.4.2
isort==4.3.4
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
PasteDeploy==2.0.1
plaster==1.0
plaster-pastedeploy==0.6
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pycparser==2.19
Pygments==2.3.1
pylint==2.2.2
pyramid==1.10.1
pyramid-debugtoolbar==4.5
pyramid-mako==1.0.2
pyramid-retry==1.0
pyramid-tm==2.2.1
python-dateutil==2.7.5
python-editor==1.0.3
repoze.lru==0.7
six==1.12.0
SQLAlchemy==1.2.15
transaction==2.4.0
translationstring==1.3
typed-ast==1.1.1
venusian==1.1.0
waitress==1.1.0
WebOb==1.8.4
wrapt==1.10.11
WTForms==2.2.1
zope.deprecation==4.4.0
zope.interface==4.6.0
zope.sqlalchemy==1.0
